web/backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "flowcase-web-backend"
version = "0.1.0"
description = "FastAPI orchestrator for the Flowcase web app (agent chat + admin)"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Marius Hernes" },
]
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "pydantic[email]>=2.7.0",
    "pydantic-settings>=2.4.0",
    "python-dotenv>=1.0.0",

    # Auth — local user/password + JWT
    "bcrypt>=4.2.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",

    # Storage — Cosmos DB (SQL / NoSQL API)
    "azure-cosmos>=4.7.0",
    "azure-identity>=1.17.0",
    "aiohttp>=3.10.0",

    # LLM — Azure OpenAI via the OpenAI SDK
    "openai>=1.40.0",

    # MCP client over streamable HTTP
    "mcp>=1.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
flowcase-web = "flowcase_web.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/flowcase_web"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
